rewrite bu to be more intuitive
The code has been refactored completely. bu is now called as such:

  bu [-l|-h] <in|out> <[-]item>

The direction must be explicitly stated by the user, preventing some
mistakes. Items are referenced by their aliases. Multiple aliases may be
specified at a time. A hyphen in front of an alias designates that alias
to be subject to deletion/full sync. The 'all' alias is a set of all
aliases. The '-l' option lists the item aliases. The '-h' option
displays a usage statement.