Make makefile more usable.
A check target has been added to see if there are differences between
the installed and local scripts.

The script names are now stored in a variable. Each target loops through
the list of script names. The install target provides more useful
information about what it's doing. When it's updating a script, it says
'updating $script' and when it is adding a script, it says 'adding
$script'.

Targets are now phony and silent.