Add `qln`.
This script manages a single directory containing a bunch of links. This
directory is "$HOME/ql" by default.

Adding links is quick. All you have to do is use 'qln' like so: `qln
file`. That will create a new link named "file" to $(realpath "file") in
`qln`'s managed directory.

If a link named "file" already exists, `qln file` will print the full
path of the file that the "file" link points to. If the "-f" flag is
passed to `qln`, all of the links after it will be created regardless
of whether or not they already exist.